+++ src/Helpers/helpers.php
@@ -587,7 +587,7 @@ if (! function_exists('constructInputs')) {
             
             //construct label
             if ($trans == true) {
-                $label = trans($input['label']);
+                $label = trans("pakka::".$input['label']);
             } else {
                 $label = $input['label'];
             }
@@ -820,20 +820,16 @@ if (! function_exists('checkAcces')) {
     {
         if (isset(auth()->user()->role)) {
             $settings = Session::get('settings');
-        
-            if (! empty($settings) && array_key_exists($setting, $settings)) {
+            if (!empty($settings) && array_key_exists($setting, $settings)) {
                 switch (true) {
                     case auth()->user()->role == 10:
                         return true;
-
                         break;
                     case auth()->user()->role !== 10 && $settings[$setting] == 1:
                         return true;
-
                         break;
                     default:
                         return false;
-
                         break;
                 }
             } else {
@@ -2056,7 +2052,6 @@ if (! function_exists('parseContent')) {
                     $empty = true;
                 }
             } else {
-                dd(App::getLocale());
                 $value = trans("pakka::app.insert_here");
                 $empty = true;
             }