+++ src/Http/Controllers/Auth/LoginController.php
@@ -8,6 +8,7 @@ use Illuminate\Foundation\Auth\AuthenticatesUsers;
 use Illuminate\Http\Request;
 use Session;
 use TheRealJanJanssens\Pakka\Http\Controllers\Controller;
+use Illuminate\Validation\ValidationException;
 
 class LoginController extends Controller
 {
@@ -48,9 +49,25 @@ class LoginController extends Controller
 
     protected function authenticated(Request $request, $user)
     {
+        //dd($user->role);
         //This is getting used as a fallback on the empty Auth::user() in helpers.php
         Session::put('auth.id', $user->id);
         Session::put('auth.email', $user->email);
         Session::put('auth.role', $user->role);
+    }
+
+    /**
+     * Get the failed login response instance.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Symfony\Component\HttpFoundation\Response
+     *
+     * @throws \Illuminate\Validation\ValidationException
+     */
+    protected function sendFailedLoginResponse(Request $request)
+    {
+        throw ValidationException::withMessages([
+            $this->username() => [trans('pakka::auth.failed')],
+        ]);
     }
 }